adding code to chroot to mounted disk image and initialize filesystem
I propose that we revert the changes in lkl_mount_dev to generate a mount point
under /mnt, and then in fdinit we use a static variable to make it so that the
first block device is chrooted to and then (if writable) initialized (creating
dev, tmpfs, etc) after mounting. Or, we might want to give them predictable
names based on order in the rexec command line (e.g., /mnt/001, /mnt/002,
etc...).